requirements.txt
apache-airflow[celery,postgres]==1.10.14
singleton-decorator==1.0.0
SQLAlchemy==1.3.15

